+++ apps/web/components/settings/workflow-sync-section.tsx
@@ -6,9 +6,10 @@ import { WorkflowSyncDialog } from "@/components/settings/workflow-sync-dialog";
 import { WorkflowSyncStatusCard } from "@/components/settings/workflow-sync-status-banner";
 import { useWorkflowSync } from "@/hooks/domains/settings/use-workflow-sync";
 
-// WorkflowSyncButton is the entry point rendered in the page header; the
-// dialog open state lives with the caller so button and section can sit in
-// different parts of the layout.
+// WorkflowSyncButton is the GitHub Sync entry point, rendered alongside the
+// other workflow actions (Export / Import / Add). The dialog open state lives
+// with the caller so button and section can sit in different parts of the
+// layout.
 export function WorkflowSyncButton({ onClick }: { onClick: () => void }) {
   return (
     <Button
@@ -31,11 +32,11 @@ type WorkflowSyncSectionProps = {
   onDialogOpenChange: (open: boolean) => void;
 };
 
-// WorkflowSyncSection renders the GitHub-sync state on the workspace
-// Workflows settings page: the configuration dialog (opened via
-// WorkflowSyncButton in the header) and — once a sync is configured — a
-// compact status card showing what is syncing and how the last attempt went.
-// Unconfigured workspaces render nothing visible.
+// WorkflowSyncSection renders the GitHub-sync state inside the Workflows
+// settings section: the configuration dialog (opened via WorkflowSyncButton
+// in the section's action row) and — once a sync is configured — a compact
+// status card above the workflow list showing what is syncing and how the
+// last attempt went. Unconfigured workspaces render nothing visible.
 export function WorkflowSyncSection({
   workspaceId,
   dialogOpen,
@@ -46,11 +47,13 @@ export function WorkflowSyncSection({
   return (
     <>
       {sync.config && (
-        <WorkflowSyncStatusCard
-          config={sync.config}
-          syncing={sync.syncing}
-          onSyncNow={sync.handleSyncNow}
-        />
+        <div className="mb-4" data-testid="workflow-sync-section">
+          <WorkflowSyncStatusCard
+            config={sync.config}
+            syncing={sync.syncing}
+            onSyncNow={sync.handleSyncNow}
+          />
+        </div>
       )}
       <WorkflowSyncDialog open={dialogOpen} onOpenChange={onDialogOpenChange} sync={sync} />
     </>
